requirements.txt
asgiref
black
certifi
cffi
charset-normalizer
click
coreapi
coreschema
cryptography
defusedxml
Django
django-cors-headers
django-filter
django-templated-mail
djangorestframework
djangorestframework-simplejwt
django-phonenumber-field==6.0.0
djoser
drf-nested-routers
drf-yasg
idna
inflection
isort
itypes
Jinja2
MarkupSafe
mypy-extensions
oauthlib
packaging
pathspec
phonenumbers==8.12.41
Pillow
platformdirs
psycopg2
pycparser
PyJWT
pyparsing
python-dotenv
python3-openid
pytz
requests
requests-oauthlib
ruamel.yaml
ruamel.yaml.clib
six
social-auth-app-django
social-auth-core
sqlparse
tomli
uritemplate
urllib3
drf-spectacular
